FYP-PENAMBAHBAIKAN KUALITI IMEJ DALAM AIR MENGGUNAKAN GABUNGAN TEKNIK PRIOR SALURAN GELAP DAN RANGKAIAN NEURAL KONVOLUSI/requirements.txt
numpy
opencv-python
torch
torchvision
Pillow
scikit-image
matplotlib
pandas
colormath
PyQt5
scikit-learn
tqdm
